+++ python/code_saturne/gui/case/StartRestartView.py
@@ -4,7 +4,7 @@
 
 # This file is part of code_saturne, a general-purpose CFD tool.
 #
-# Copyright (C) 1998-2024 EDF S.A.
+# Copyright (C) 1998-2026 EDF S.A.
 #
 # This program is free software; you can redistribute it and/or modify it under
 # the terms of the GNU General Public License as published by the Free Software
@@ -234,17 +234,25 @@ class StartRestartView(QWidget, Ui_StartRestartForm):
         self.radioButtonNo.clicked.connect(self.slotStartRestart)
         self.radioButtonAuto.clicked.connect(self.slotStartRestart)
         self.toolButton.pressed.connect(self.slotSearchRestartDirectory)
-        self.toolButtonRestartMesh.pressed.connect(self.slotSearchRestartMesh)
         self.checkBox.clicked.connect(self.slotFrozenField)
         self.toolButtonAdvanced.pressed.connect(self.slotAdvancedOptions)
-        self.checkBoxRestartMesh.stateChanged.connect(self.slotRestartMesh)
+
+        self.__comboModelMesh = ComboModel(self.comboBoxMesh, 4, 1)
+        self.__comboModelMesh.addItem(self.tr("Automatic (unmodified if present)"),
+                                      "automatic")
+        self.__comboModelMesh.addItem(self.tr("Unmodified (disable preprocessing)"),
+                                      "unmodified")
+        self.__comboModelMesh.addItem(self.tr("Different mesh (interpolate)"),
+                                      "different_mesh")
+        self.__comboModelMesh.addItem(self.tr("Rebuild same mesh"),
+                                      "same_mesh_preprocess")
+        self.comboBoxMesh.activated[str].connect(self.slotRestartMesh)
 
         self.model = StartRestartModel(self.case)
 
         # Widget initialization
 
         self.restart_path = self.model.getRestartPath()
-        self.restart_mesh_path = self.model.getRestartMeshPath()
 
         if self.restart_path:
             if self.restart_path == '*':
@@ -275,7 +283,7 @@ class StartRestartView(QWidget, Ui_StartRestartForm):
             self.checkBox.setChecked(False)
 
         self.updateRestartTimes()
-        self.updateRestartMeshView()
+        self.updateRestartMeshBehavior()
 
         self.case.undoStartGlobal()
 
@@ -335,20 +343,12 @@ class StartRestartView(QWidget, Ui_StartRestartForm):
             self.lineEditTime.hide()
 
 
-    def updateRestartMeshView(self):
+    def updateRestartMeshBehavior(self):
         """
-        Upate restart mesh path view
+        Update information on restart mesh behavior
         """
-        if self.restart_mesh_path:
-            self.checkBoxRestartMesh.setChecked(True)
-            self.lineEditRestartMesh.setText(self.restart_mesh_path)
-            self.lineEditRestartMesh.show()
-            self.toolButtonRestartMesh.show()
-        else:
-            self.checkBoxRestartMesh.setChecked(False)
-            self.lineEditRestartMesh.setText("")
-            self.lineEditRestartMesh.hide()
-            self.toolButtonRestartMesh.hide()
+        v = self.model.getRestartMeshBehavior()
+        self.__comboModelMesh.setItem(str_model=v)
 
 
     @pyqtSlot()
@@ -402,55 +402,6 @@ class StartRestartView(QWidget, Ui_StartRestartForm):
 
 
     @pyqtSlot()
-    def slotSearchRestartMesh(self):
-        """
-        Search restart mesh (file) in list of directories
-        """
-        title    = self.tr("Select checkpoint/restart mesh_input/output")
-
-        default = None
-        l_restart_dirs = []
-        for d in [os.path.join(os.path.split(self.case['case_path'])[0],
-                               'RESU_COUPLING'),
-                  os.path.join(self.case['case_path'], 'RESU')]:
-            if os.path.isdir(d):
-                l_restart_dirs.append(QUrl.fromLocalFile(d))
-                if not default:
-                    default = d
-
-        if not default:
-            default = self.case['case_path']
-
-        options  = QFileDialog.Option.DontUseNativeDialog | QFileDialog.Option.ReadOnly
-
-        dialog = QFileDialog()
-        dialog.setWindowTitle(title)
-        dialog.setDirectory(default)
-
-        dialog.setOptions(options)
-        dialog.setSidebarUrls(l_restart_dirs)
-        dialog.setFileMode(QFileDialog.FileMode.ExistingFile)
-
-        name_filter = str(self.tr("Imported or preprocessed mesh (mesh_input mesh_output *.csm)"))
-        dialog.setNameFilter(name_filter)
-
-        dialog.setLabelText(QFileDialog.DialogLabel.Accept, str(self.tr("Select")))
-
-        if dialog.exec_() == 1:
-
-            s = dialog.selectedFiles()
-
-            path = str(s[0])
-            path = os.path.abspath(path)
-
-            self.restart_mesh_path = RelOrAbsPath(path, self.case['case_path'])
-            self.model.setRestartMeshPath(self.restart_mesh_path)
-            self.lineEditRestartMesh.setText(self.restart_mesh_path)
-
-            log.debug("slotSearchRestartDirectory-> %s" % self.restart_mesh_path)
-
-
-    @pyqtSlot()
     def slotStartRestart(self):
         """
         Handle restart.
@@ -480,33 +431,26 @@ class StartRestartView(QWidget, Ui_StartRestartForm):
             self.frameRestart.show()
         else:
             self.model.setRestartPath(None)
-            self.model.setRestartMeshPath(None)
             self.model.setFrozenField("off")
+            self.model.setRestartMeshBehavior("unmodified")
+            self.updateRestartMeshBehavior()
             self.radioButtonYes.setChecked(False)
             self.radioButtonNo.setChecked(True)
-            self.checkBoxRestartMesh.setChecked(False)
             self.frameRestart.hide()
             self.lineEdit.setText("")
             self.updateRestartTimes()
 
-        self.updateRestartMeshView()
 
-
-    @pyqtSlot()
-    def slotRestartMesh(self):
+    @pyqtSlot(str)
+    def slotRestartMesh(self, text):
         """
-        Input different restart mesh.
+        Define restart mesh behavior.
         """
-        if self.checkBoxRestartMesh.isChecked():
-            if not self.restart_mesh_path:
-                self.slotSearchRestartMesh()
+        v = self.__comboModelMesh.dicoV2M[str(text)]
+        if v == self.model.getRestartMeshBehavior():
+            return
 
-        else:
-            self.restart_mesh_path = None
-
-        self.model.setRestartMeshPath(self.restart_mesh_path)
-
-        self.updateRestartMeshView()
+        self.model.setRestartMeshBehavior(v)
 
 
     @pyqtSlot()